+++ client/src/assetManager/contextMenu.vue
@@ -10,7 +10,6 @@ import Vue from "vue";
 
 import Component from "vue-class-component";
 
-import AssetManager from "@/assetManager/manager.vue";
 import ContextMenu from "@/core/components/contextmenu.vue";
 import ConfirmDialog from "@/core/components/modals/confirm.vue";
 import Prompt from "@/core/components/modals/prompt.vue";
@@ -69,7 +68,8 @@ export default class AssetContextMenu extends Vue {
                     if (result) {
                         for (const sel of assetStore.selected) {
                             socket.emit("Asset.Remove", sel);
-                            if (assetStore.isFile(sel)) assetStore.files.splice(assetStore.files.indexOf(sel), 1);
+                            if (assetStore.files.includes(sel))
+                                assetStore.files.splice(assetStore.files.indexOf(sel), 1);
                             else assetStore.folders.splice(assetStore.folders.indexOf(sel), 1);
                         }
                         assetStore.clearSelected();