+++ CHANGELOG.md
@@ -31,6 +31,7 @@ All notable changes to this project will be documented in this file.
 - Active location not being remembered by server.
 - Notes not getting cleared on location change.
 - AssetManager shift selection causing double selections.
+- AssetManager issues with (re)moving files.
 
 ## [0.13.3] - 2019-01-19